+++ src/code42cli/options.py
@@ -3,9 +3,22 @@ import click
 from code42cli.cmds.search.enums import ServerProtocol
 from code42cli.errors import Code42CLIError
 from code42cli.output_formats import OutputFormat
+from code42cli.output_formats import SendToFileEventsOutputFormat
 from code42cli.profile import get_profile
 from code42cli.sdk_client import create_sdk
 
+BEGIN_OPTION_HELP_MESSAGE = (
+    "The beginning of the date range in which to look for {}, can be a date/time in "
+    "yyyy-MM-dd (UTC) or yyyy-MM-dd HH:MM:SS (UTC+24-hr time) format where the 'time' "
+    "portion of the string can be partial (e.g. '2020-01-01 12' or '2020-01-01 01:15') "
+    "or a short value representing days (30d), hours (24h) or minutes (15m) from current "
+    "time."
+)
+
+END_OPTION_HELP_MESSAGE = (
+    "The end of the date range in which to look for {}, argument format options are "
+    "the same as `--begin`."
+)
 
 yes_option = click.option(
     "-y",
@@ -23,6 +36,23 @@ format_option = click.option(
     help="The output format of the result. Defaults to table format.",
     default=OutputFormat.TABLE,
 )
+
+
+def begin_option(f, search_term, **kwargs):
+    start_time = click.option(
+        "-b", "--begin", help=BEGIN_OPTION_HELP_MESSAGE.format(search_term), **kwargs
+    )
+
+    f = start_time(f)
+    return f
+
+
+def end_option(f, search_term, **kwargs):
+    end_time = click.option(
+        "-e", "--end", help=END_OPTION_HELP_MESSAGE.format(search_term), **kwargs
+    )
+    f = end_time(f)
+    return f
 
 
 class CLIState:
@@ -120,3 +150,12 @@ def server_options(f):
     f = hostname_arg(f)
     f = protocol_option(f)
     return f
+
+
+send_to_format_options = click.option(
+    "-f",
+    "--format",
+    type=click.Choice(SendToFileEventsOutputFormat(), case_sensitive=False),
+    help="The output format of the result. Defaults to json format.",
+    default=SendToFileEventsOutputFormat.RAW,
+)